[skip ci] • ego_model.py: fixed gain (to be 1)   now identical to pytorch for lr==0

+++ Scripts/Models (Under Development)/EGO/Using EMComposition/Coffee Shop World/younes/data/dataset.py
@@ -215,4 +215,4 @@ def one_hot_encode(labels, num_classes):
     """
     One hot encode labels and convert to tensor.
     """
-    return torch.tensor((np.arange(num_classes) == labels[..., None]).astype(float), dtype=torch.float32)
+    return torch.tensor((np.arange(num_classes) == labels[..., None]).astype(float), dtype=torch.float64)
